fixup! fixup! runtime/secrets: add package for consolidated secret handling
refactor to share TLS certificate handling logic
